Rename icc package to notify
