+++ internal/notify/channel_id_test.go
@@ -1,4 +1,4 @@
-package icc
+package notify
 
 import "testing"
