+++ internal/notify/http_test.go
@@ -1,4 +1,4 @@
-package icc_test
+package notify_test
 
 import (
 	"errors"
@@ -7,9 +7,9 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/OpenSlides/openslides-icc-service/internal/icc"
 	"github.com/OpenSlides/openslides-icc-service/internal/iccerror"
 	"github.com/OpenSlides/openslides-icc-service/internal/icctest"
+	"github.com/OpenSlides/openslides-icc-service/internal/notify"
 )
 
 func TestHandleReceive(t *testing.T) {
@@ -21,7 +21,7 @@ func TestHandleReceive(t *testing.T) {
 			expectedMessage: "my answer",
 		}
 		mux := http.NewServeMux()
-		icc.HandleReceive(mux, &receiver, &auther)
+		notify.HandleReceive(mux, &receiver, &auther)
 		resp := httptest.NewRecorder()
 
 		mux.ServeHTTP(resp, httptest.NewRequest("GET", url, nil))
@@ -47,7 +47,7 @@ func TestHandleReceive(t *testing.T) {
 			UserID: 1,
 		}
 		mux := http.NewServeMux()
-		icc.HandleReceive(mux, &receiver, &auther)
+		notify.HandleReceive(mux, &receiver, &auther)
 		resp := httptest.NewRecorder()
 
 		mux.ServeHTTP(resp, httptest.NewRequest("GET", url, nil))
@@ -73,7 +73,7 @@ func TestHandleReceive(t *testing.T) {
 			UserID: 1,
 		}
 		mux := http.NewServeMux()
-		icc.HandleReceive(mux, &receiver, &auther)
+		notify.HandleReceive(mux, &receiver, &auther)
 		resp := httptest.NewRecorder()
 
 		mux.ServeHTTP(resp, httptest.NewRequest("GET", url+"?meeting_id=5", nil))
@@ -104,7 +104,7 @@ func TestHandleReceive(t *testing.T) {
 			UserID: 1,
 		}
 		mux := http.NewServeMux()
-		icc.HandleReceive(mux, &receiver, &auther)
+		notify.HandleReceive(mux, &receiver, &auther)
 		resp := httptest.NewRecorder()
 
 		mux.ServeHTTP(resp, httptest.NewRequest("GET", url, nil))
@@ -127,7 +127,7 @@ func TestHandleReceive(t *testing.T) {
 			UserID: 1,
 		}
 		mux := http.NewServeMux()
-		icc.HandleReceive(mux, &receiver, &auther)
+		notify.HandleReceive(mux, &receiver, &auther)
 		resp := httptest.NewRecorder()
 
 		mux.ServeHTTP(resp, httptest.NewRequest("GET", url, nil))
@@ -149,7 +149,7 @@ func TestHandleSend(t *testing.T) {
 		auther := icctest.AutherStub{}
 		sender := senderStub{}
 		mux := http.NewServeMux()
-		icc.HandleSend(mux, &sender, &auther)
+		notify.HandleSend(mux, &sender, &auther)
 		resp := httptest.NewRecorder()
 
 		mux.ServeHTTP(resp, httptest.NewRequest("GET", url, nil))
@@ -173,7 +173,7 @@ func TestHandleSend(t *testing.T) {
 		}
 		sender := senderStub{}
 		mux := http.NewServeMux()
-		icc.HandleSend(mux, &sender, &auther)
+		notify.HandleSend(mux, &sender, &auther)
 		resp := httptest.NewRecorder()
 
 		mux.ServeHTTP(resp, httptest.NewRequest("GET", url, nil))
@@ -200,7 +200,7 @@ func TestHandleSend(t *testing.T) {
 			UserID: 1,
 		}
 		mux := http.NewServeMux()
-		icc.HandleSend(mux, &sender, &auther)
+		notify.HandleSend(mux, &sender, &auther)
 		resp := httptest.NewRecorder()
 
 		mux.ServeHTTP(resp, httptest.NewRequest("GET", url, nil))
